fixed the size of timestamps
git-svn-id: svn://svn.minisip.org/minisip/trunk@943 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ minisip/video/ImageHandler.h
@@ -55,7 +55,7 @@ typedef struct MImage{
 	int linesize[4];
 	uint32_t ssrc;       
 	uint32_t chroma;
-	uint32_t mTime;      /* TimeStamp in µsec */
+	uint64_t mTime;      /* TimeStamp in msec */
 	void * privateData;  /* Can be used by the ImageHandlers */
 } MImage;
